Move client code to unified render & fix unix build issues
